Standard image format is now in [0,1] instead of [0,255].
To eliminate the need to consider the max color/gray value as meta-data with
the image, images are stored in a standard [0,1] range. Hence, only Netpbm and
PNG exporting functions need consider the max color/gray value. The k
parameter was removed from all functions that were transformative in nature.
With this change, there is motivation to eliminate the redundant Netpbm class.

It should be noted that there is an additional piece of meta-data that is
assumed in the notion of a "standard image." That is the color space of the
image which is assumed to be RGB. Thus, when manipulating color spaces, the
user is responsible for managing which color space an image is currently in and
converting the color space back to RGB when exporting.