Remove k from read_netpbm return statement.
The reasoning for this is that the [0,1] continous form of the image does not
lose any informaton and it is not vital that the user knows the k value the
Netpbm file happened to use because any k value could be used when exporting to
Netpbm again. Furthermore, it is especially worth getting rid this information
to gain the benefit of a general image read function.