Created ResizeFilter namedtuple and added docs to ResizeFilterName.
The ResizeFilter namedtuple provides a place to nicely define the components
that make up an image resize filter. The rescale function now takes a single
filter argument which can be a ResizeFilter (reducing confusion with two
arguments to specify custom filters) or a ResizeFilterName which is a member of
the new enum ResizeFilterName. The documentation of this enum was also improved
by moving the links to ImageMagick implementation there.