Added TODO with potential change to _continuous and _discretize.
The current implementation of _continuous, $x/k$ maps 0 to 0 and k to 1 which
feels desirable. However, I beleive _discretize should be the inverse of
_continuous which results in the current implementation of _discretize being
$ceil(x*k - 0.5)$ which has "bins" of differing widths at the extremes (half
the width of the internal bins). This does not feel right. The TODO notes
describe a counter-proposal in which _discretize maps to constant-width bins
but the inverse means that _continuous no longer maps 0 and k to 0 and 1
respectivley. This also feels wrong.